+++ tests/base/test_alarm.py
@@ -34,26 +34,11 @@ from lsst.ts import watcher
 STD_TIMEOUT = 2  # seconds
 
 
-class AsyncCallback:
-    """Callback functor with a future that is set done when called."""
-
-    def __init__(self):
-        self._future = asyncio.Future()
-
-    def __call__(self, alarm):
-        if not self._future.done():
-            self._future.set_result(alarm)
-
-    async def next(self, timeout):
-        await asyncio.wait_for(self._future, timeout=timeout)
-        self._future = asyncio.Future()
-
-
 class AlarmTestCase(asynctest.TestCase):
     # NOTE: almost all test methods must be async, even with no visible
     # async code, so that Alarm has an event loop.
     def setUp(self):
-        self.ncalls = 0
+        self.callback_queue = asyncio.Queue()
         self._alarms = []  # Alarms to close in tearDown.
 
     def tearDown(self):
@@ -61,10 +46,38 @@ class AlarmTestCase(asynctest.TestCase):
             alarm.close()
 
     def callback(self, alarm):
-        self.ncalls += 1
+        self.callback_queue.put_nowait(alarm)
 
-    def alarm_iter(self, callback, name="test.alarm"):
-        """Return an iterator over alarms with all allowed values of
+    async def next_queued_alarm(self, expected_alarm=None, timeout=STD_TIMEOUT):
+        """Get the next alarm queued by `callback`.
+
+        Parameters
+        ----------
+        expected_alarm : `lsst.ts.watcher.base.Alarm` or `None`, optional
+            The alarm that should be returned.
+            If None (the default) then this is not checked.
+        timeout : `float`
+            Time limit for waiting (seconds).
+        """
+        alarm = await asyncio.wait_for(self.callback_queue.get(), timeout=STD_TIMEOUT)
+        if expected_alarm is not None:
+            self.assertIs(alarm, expected_alarm)
+        return alarm
+
+    @property
+    def ncalls(self):
+        return self.callback_queue.qsize()
+
+    def alarm_iter(
+        self,
+        callback,
+        name="test.alarm",
+        auto_acknowledge_delay=0,
+        auto_unacknowledge_delay=0,
+        escalate_to="",
+        escalate_delay=0,
+    ):
+        """An iterator over alarms with all allowed values of
         severity and max_severity.
 
         Parameters
@@ -72,21 +85,45 @@ class AlarmTestCase(asynctest.TestCase):
         callback : callable or `None`
             Callback function; must take one argument: an alarm.
             None for no callback function.
-        name : `str` (optional)
+        name : `str`, optional
             Name of alarm.
+        auto_acknowledge_delay : `float`, optional
+            Delay (in seconds) before a stale alarm is automatically
+            acknowledged, or 0 for no automatic acknowledgement.
+        auto_unacknowledge_delay : `float`, optional
+            Delay (in seconds) before an acknowledged alarm is automatically
+            unacknowledged, or 0 for no automatic unacknowledgement.
+        escalate_to : `str`, optional
+            Who or what to escalate the alarm to.
+            If "" (the default) the alarm is not escalated.
+        escalate_delay : `float`, optional
+            Delay before escalating a critical unacknowledged alarm (sec).
+            If 0 (the default) the alarm is not escalated.
 
         Notes
         -----
-        Does not affect self.ncalls.
+        The callback function is not called by this iterator,
+        so this iterator does not change `self.ncalls`.
         """
 
         def _alarm_iter_impl():
-            # Generate alarms without a callback, then set the callback
-            # at the very end, after setting the severity.
+            """An iterator over alarms with all allowed values of
+            severity and max_severity and no callback function.
+
+            Assign the callback later, if desired.
+            """
+            alarm_kwargs = dict(
+                name=name,
+                auto_acknowledge_delay=auto_acknowledge_delay,
+                auto_unacknowledge_delay=auto_unacknowledge_delay,
+                escalate_to=escalate_to,
+                escalate_delay=escalate_delay,
+            )
+
             severities = list(AlarmSeverity)
-            yield self.make_alarm(name=name)
+            yield self.make_alarm(**alarm_kwargs)
             for i, severity in enumerate(severities[1:]):
-                alarm = self.make_alarm(name=name)
+                alarm = self.make_alarm(**alarm_kwargs)
                 reason = f"alarm_iter set severity={severity}"
                 updated = alarm.set_severity(severity=severity, reason=reason)
                 self.assertTrue(updated)
@@ -94,7 +131,7 @@ class AlarmTestCase(asynctest.TestCase):
 
             for i, max_severity in enumerate(severities[1:]):
                 for severity in severities[0 : i + 1]:
-                    alarm = self.make_alarm(name=name)
+                    alarm = self.make_alarm(**alarm_kwargs)
                     updated = alarm.set_severity(severity=max_severity, reason=reason)
                     self.assertTrue(updated)
                     reason = f"alarm_iter set severity to {severity} after setting it to {max_severity}"
@@ -103,11 +140,29 @@ class AlarmTestCase(asynctest.TestCase):
                     yield alarm
 
         for alarm in _alarm_iter_impl():
-            alarm.configure(callback=callback)
+            alarm.callback = callback
             yield alarm
 
+    def copy_alarm(self, alarm):
+        """Return a shallow copy of an alarm and schedule it to be closed.
+
+        Parameters
+        ----------
+        alarm : `lsst.ts.watcher.base.Alarm`
+            The alarm to copy.
+        """
+        alarm_copy = copy.copy(alarm)
+        self._alarms.append(alarm_copy)
+        return alarm_copy
+
     def make_alarm(
-        self, name, callback=None, auto_acknowledge_delay=0, auto_unacknowledge_delay=0
+        self,
+        name,
+        callback=None,
+        auto_acknowledge_delay=0,
+        auto_unacknowledge_delay=0,
+        escalate_to="",
+        escalate_delay=0,
     ):
         """Make an alarm and keep a reference so tearDown can close it.
 
@@ -115,21 +170,29 @@ class AlarmTestCase(asynctest.TestCase):
         ----------
         name : `str`
             Name of alarm.
-        callback : callable or `None` (optional)
+        callback : callable or `None`, optional
             Callback function; must take one argument: an alarm.
             None for no callback function.
-        auto_acknowledge_delay : `float` (optional)
+        auto_acknowledge_delay : `float`, optional
             Delay (in seconds) before a stale alarm is automatically
             acknowledged, or 0 for no automatic acknowledgement.
-        auto_unacknowledge_delay : `float` (optional)
+        auto_unacknowledge_delay : `float`, optional
             Delay (in seconds) before an acknowledged alarm is automatically
             unacknowledged, or 0 for no automatic unacknowledgement.
+        escalate_to : `str`, optional
+            Who or what to escalate the alarm to.
+            If "" (the default) the alarm is not escalated.
+        escalate_delay : `float`, optional
+            Delay before escalating a critical unacknowledged alarm (sec).
+            If 0 (the default) the alarm is not escalated.
         """
         alarm = watcher.base.Alarm(name=name)
         alarm.configure(
             callback=callback,
             auto_acknowledge_delay=auto_acknowledge_delay,
             auto_unacknowledge_delay=auto_unacknowledge_delay,
+            escalate_to=escalate_to,
+            escalate_delay=escalate_delay,
         )
         self._alarms.append(alarm)
         return alarm
@@ -313,89 +376,115 @@ class AlarmTestCase(asynctest.TestCase):
         self.assertTrue(alarm.auto_unacknowledge_task.done())
 
     async def test_alarm_iter(self):
-        name = "stella"
-        severity_set = set()
-        nitems = 0
         nseverities = len(AlarmSeverity)
-        predicted_nitems = nseverities * (nseverities + 1) // 2
-        for alarm in self.alarm_iter(name=name, callback=self.callback):
-            nitems += 1
-            self.assertEqual(self.ncalls, 0)
-            self.assertEqual(alarm.name, name)
-            self.assertGreaterEqual(alarm.max_severity, alarm.severity)
-            self.assertFalse(alarm.acknowledged)
-            self.assertEqual(alarm.acknowledged_by, "")
-            self.assertFalse(alarm.escalated)
-            self.assertEqual(alarm.escalate_to, "")
-            self.assertEqual(alarm.muted_severity, AlarmSeverity.NONE)
-            self.assertEqual(alarm.muted_by, "")
-            if nitems == 1:
-                # first state is NONE, alarm is nominal
-                self.assertTrue(alarm.nominal)
-                self.assertEqual(alarm.timestamp_severity_oldest, 0)
-                self.assertEqual(alarm.timestamp_severity_newest, 0)
-                self.assertEqual(alarm.timestamp_max_severity, 0)
-            else:
-                self.assertFalse(alarm.nominal)
-                self.assertGreater(alarm.timestamp_severity_oldest, 0)
-                self.assertGreater(alarm.timestamp_severity_newest, 0)
-                self.assertGreater(alarm.timestamp_max_severity, 0)
-            self.assertEqual(alarm.timestamp_auto_acknowledge, 0)
-            self.assertGreaterEqual(
-                alarm.timestamp_severity_newest, alarm.timestamp_severity_oldest
-            )
-            self.assertEqual(alarm.timestamp_acknowledged, 0)
-            self.assertEqual(alarm.timestamp_auto_unacknowledge, 0)
-            self.assertEqual(alarm.timestamp_escalate, 0)
-            self.assertEqual(alarm.timestamp_unmute, 0)
-            severity_set.add((alarm.severity, alarm.max_severity))
-        self.assertEqual(nitems, predicted_nitems)
-        self.assertEqual(nitems, len(severity_set))
+
+        # Default arguments to self.alarm_iter
+        default_kwargs = dict(
+            name="test.alarm",
+            auto_acknowledge_delay=0,
+            auto_unacknowledge_delay=0,
+            escalate_to="",
+            escalate_delay=0,
+        )
+        for callback, kwargs in itertools.product(
+            (None, self.callback),
+            (
+                dict(),
+                dict(name="foo"),
+                dict(auto_acknowledge_delay=1),
+                dict(auto_unacknowledge_delay=2),
+                # Set both escalate_to and escalate_delay
+                # to enable escalation.
+                dict(escalate_to="chaos", escalate_delay=3),
+            ),
+        ):
+
+            def assert_expected(alarm):
+                """Assert that all alarm fields that can be specified
+                when calling `alarm_iter` have the expected value.
+                """
+                for argname, default_value in default_kwargs.items():
+                    expected_value = kwargs.get(argname, default_value)
+                    self.assertEqual(
+                        getattr(alarm, argname), expected_value, msg=argname
+                    )
+
+            severity_set = set()
+            nitems = 0
+            predicted_nitems = nseverities * (nseverities + 1) // 2
+            for alarm in self.alarm_iter(callback=callback, **kwargs):
+                nitems += 1
+                self.assertEqual(self.ncalls, 0)
+                self.assertEqual(alarm.callback, callback)
+                assert_expected(alarm)
+
+                self.assertGreaterEqual(alarm.max_severity, alarm.severity)
+                self.assertFalse(alarm.acknowledged)
+                self.assertEqual(alarm.acknowledged_by, "")
+                self.assertFalse(alarm.escalated)
+                self.assertEqual(alarm.muted_severity, AlarmSeverity.NONE)
+                self.assertEqual(alarm.muted_by, "")
+                if nitems == 1:
+                    # first state is NONE, alarm is nominal
+                    self.assertTrue(alarm.nominal)
+                    self.assertEqual(alarm.timestamp_severity_oldest, 0)
+                    self.assertEqual(alarm.timestamp_severity_newest, 0)
+                    self.assertEqual(alarm.timestamp_max_severity, 0)
+                else:
+                    self.assertFalse(alarm.nominal)
+                    self.assertGreater(alarm.timestamp_severity_oldest, 0)
+                    self.assertGreater(alarm.timestamp_severity_newest, 0)
+                    self.assertGreater(alarm.timestamp_max_severity, 0)
+
+                self.assertEqual(alarm.timestamp_acknowledged, 0)
+                self.assertEqual(alarm.timestamp_auto_unacknowledge, 0)
+
+                stale_alarm = (
+                    alarm.severity == AlarmSeverity.NONE
+                    and alarm.max_severity > AlarmSeverity.NONE
+                )
+                if stale_alarm and alarm.auto_acknowledge_delay > 0:
+                    self.assertGreater(alarm.timestamp_auto_acknowledge, 0)
+                else:
+                    self.assertEqual(alarm.timestamp_auto_acknowledge, 0)
+                self.assertGreaterEqual(
+                    alarm.timestamp_severity_newest, alarm.timestamp_severity_oldest
+                )
+
+                auto_escalate = (
+                    alarm.max_severity == AlarmSeverity.CRITICAL
+                    and alarm.escalate_delay > 0
+                    and alarm.escalate_to != ""
+                )
+                if auto_escalate:
+                    self.assertGreater(alarm.timestamp_escalate, 0)
+                else:
+                    self.assertEqual(alarm.timestamp_escalate, 0)
+                self.assertEqual(alarm.timestamp_unmute, 0)
+                severity_set.add((alarm.severity, alarm.max_severity))
+            self.assertEqual(nitems, predicted_nitems)
+            self.assertEqual(nitems, len(severity_set))
 
     async def test_equality(self):
         """Test __eq__ and __ne__
 
         This is a rather crude test in that it sets fields to
         invalid values.
-
-        Note: we do not close the alarm copies because there is no need:
-        no tasks are started.
         """
         alarm0 = self.make_alarm(name="foo", callback=self.callback)
-        alarm = copy.copy(alarm0)
+        alarm = self.copy_alarm(alarm0)
         self.assertTrue(alarm == alarm0)
         self.assertFalse(alarm != alarm0)
-        # All properties are ignored when testing equality
-        properties = set(["muted", "nominal"])
-        # These fields are ignored when tesing equality
-        fields_to_ignore = set(
-            [
-                "auto_acknowledge_delay",
-                "auto_unacknowledge_delay",
-                "auto_acknowledge_task",
-                "auto_unacknowledge_task",
-                "timestamp_auto_acknowledge",
-                "timestamp_auto_unacknowledge",
-                "unmute_task",
-            ]
-        )
-
-        for fieldname in dir(alarm):
+        alarm.assert_equal(alarm0)
+        for fieldname, value in vars(alarm).items():
+            if fieldname.endswith("_task"):
+                continue
             with self.subTest(fieldname=fieldname):
-                if fieldname.startswith("__"):
-                    continue
-                if fieldname in properties:
-                    continue
-                value = getattr(alarm, fieldname)
-                # ignore methods
-                if fieldname != "callback" and callable(value):
-                    continue
-                if fieldname in fields_to_ignore:
-                    continue
-                alarm = copy.copy(alarm0)
+                alarm = self.copy_alarm(alarm0)
                 setattr(alarm, fieldname, 5)
                 self.assertFalse(alarm == alarm0)
                 self.assertTrue(alarm != alarm0)
+                alarm.assert_equal(alarm0, ignore_attrs=[fieldname])
 
     async def test_constructor(self):
         name = "test_fairly_long_alarm_name"
@@ -451,7 +540,7 @@ class AlarmTestCase(asynctest.TestCase):
         """Test that decreasing severity does not decrease max_severity."""
         desired_ncalls = 0
         for alarm in self.alarm_iter(callback=self.callback):
-            alarm0 = copy.copy(alarm)
+            alarm0 = self.copy_alarm(alarm)
             for severity in reversed(list(AlarmSeverity)):
                 if severity >= alarm.severity:
                     continue
@@ -506,7 +595,7 @@ class AlarmTestCase(asynctest.TestCase):
         """Test setting the same severity multiple times."""
         desired_ncalls = 0
         for alarm in self.alarm_iter(callback=self.callback):
-            alarm0 = copy.copy(alarm)
+            alarm0 = self.copy_alarm(alarm)
 
             curr_tai = salobj.current_tai()
             reason = f"set again to {alarm.severity}"
@@ -543,56 +632,95 @@ class AlarmTestCase(asynctest.TestCase):
     async def test_acknowledge(self):
         user = "skipper"
         desired_ncalls = 0
-        for alarm0 in self.alarm_iter(callback=self.callback):
-            if alarm0.nominal:
-                continue
-            for ack_severity in AlarmSeverity:
-                alarm = copy.copy(alarm0)
+        for auto_unacknowledge_delay in (0, 1):
+            for alarm0 in self.alarm_iter(
+                callback=self.callback,
+                auto_unacknowledge_delay=auto_unacknowledge_delay,
+            ):
                 if alarm0.nominal:
-                    # ack has no effect
-                    updated = alarm.acknowledge(severity=ack_severity, user=user)
-                    self.assertFalse(updated)
-                    self.assertEqual(alarm, alarm0)
-                elif ack_severity < alarm.max_severity:
-                    # ack severity too small
-                    with self.assertRaises(ValueError):
-                        alarm.acknowledge(severity=ack_severity, user=user)
-                    self.assertEqual(alarm, alarm0)
-                else:
-                    tai1 = salobj.current_tai()
-                    updated = alarm.acknowledge(severity=ack_severity, user=user)
-                    desired_ncalls += 1
-                    self.assertTrue(updated)
-                    self.assertEqual(alarm.severity, alarm0.severity)
-                    self.assertTrue(alarm.acknowledged)
-                    self.assertEqual(alarm.acknowledged_by, user)
-                    if alarm0.severity == AlarmSeverity.NONE:
-                        # alarm is reset to nominal
-                        self.assertEqual(alarm.max_severity, AlarmSeverity.NONE)
-                        self.assertTrue(alarm.nominal)
+                    continue
+                for ack_severity in AlarmSeverity:
+                    alarm = self.copy_alarm(alarm0)
+                    self.assertEqual(
+                        alarm.auto_unacknowledge_delay, auto_unacknowledge_delay
+                    )
+                    if alarm0.nominal:
+                        # ack has no effect
+                        updated = alarm.acknowledge(severity=ack_severity, user=user)
+                        self.assertFalse(updated)
+                        self.assertEqual(alarm, alarm0)
+                    elif ack_severity < alarm.max_severity:
+                        # ack severity too small
+                        with self.assertRaises(ValueError):
+                            alarm.acknowledge(severity=ack_severity, user=user)
                     else:
-                        # alarm is still active
-                        self.assertEqual(alarm.max_severity, ack_severity)
-                        self.assertFalse(alarm.nominal)
-                    self.assertEqual(
-                        alarm.timestamp_severity_oldest,
-                        alarm0.timestamp_severity_oldest,
-                    )
-                    self.assertEqual(
-                        alarm.timestamp_severity_newest,
-                        alarm0.timestamp_severity_newest,
-                    )
-                    self.assertGreaterEqual(alarm.timestamp_max_severity, tai1)
-                    self.assertGreaterEqual(alarm.timestamp_acknowledged, tai1)
+                        tai1 = salobj.current_tai()
+                        updated = alarm.acknowledge(severity=ack_severity, user=user)
+                        desired_ncalls += 1
+                        self.assertTrue(updated)
+                        self.assertEqual(alarm.severity, alarm0.severity)
+                        self.assertTrue(alarm.acknowledged)
+                        self.assertEqual(alarm.acknowledged_by, user)
+                        self.assertEqual(alarm.timestamp_auto_acknowledge, 0)
+                        self.assertEqual(alarm.timestamp_escalate, 0)
+                        if alarm0.severity == AlarmSeverity.NONE:
+                            # alarm is reset to nominal
+                            self.assertEqual(alarm.max_severity, AlarmSeverity.NONE)
+                            self.assertTrue(alarm.nominal)
+                        else:
+                            # alarm is still active
+                            self.assertEqual(alarm.max_severity, ack_severity)
+                            self.assertFalse(alarm.nominal)
+                        self.assertEqual(
+                            alarm.timestamp_severity_oldest,
+                            alarm0.timestamp_severity_oldest,
+                        )
+                        self.assertEqual(
+                            alarm.timestamp_severity_newest,
+                            alarm0.timestamp_severity_newest,
+                        )
+                        self.assertGreaterEqual(alarm.timestamp_max_severity, tai1)
+                        self.assertGreaterEqual(alarm.timestamp_acknowledged, tai1)
+                        await asyncio.sleep(0)
 
-                    # acknowledge again; this should have no affect
-                    acked_alarm = copy.copy(alarm)
-                    user2 = "a different user"
-                    updated = alarm.acknowledge(severity=ack_severity, user=user2)
-                    self.assertFalse(updated)
-                    self.assertEqual(alarm, acked_alarm)
+                        # Check task state; sleep first to let task cancellation
+                        # happen.
+                        self.assertTrue(alarm.auto_acknowledge_task.done())
+                        self.assertTrue(alarm.escalate_task.done())
+                        if (
+                            auto_unacknowledge_delay == 0
+                            or alarm.severity == AlarmSeverity.NONE
+                        ):
+                            self.assertTrue(alarm.auto_unacknowledge_task.done())
+                        else:
+                            self.assertFalse(alarm.auto_unacknowledge_task.done())
+                        # Alarm was never muted.
+                        self.assertTrue(alarm.unmute_task.done())
 
-        self.assertEqual(self.ncalls, desired_ncalls)
+                        # Acknowledge again; this should have no affect
+                        # except possibly restarting the unack timer.
+                        restart_unack_timer = (
+                            alarm.severity > AlarmSeverity.NONE
+                            and alarm.auto_unacknowledge_delay > 0
+                        )
+                        acked_alarm = self.copy_alarm(alarm)
+                        user2 = "a different user"
+                        updated = alarm.acknowledge(severity=ack_severity, user=user2)
+                        if restart_unack_timer:
+                            self.assertTrue(updated)
+                            alarm.assert_equal(
+                                acked_alarm,
+                                ignore_attrs=["timestamp_auto_unacknowledge"],
+                            )
+                            self.assertGreater(
+                                alarm.timestamp_auto_unacknowledge,
+                                acked_alarm.timestamp_auto_unacknowledge,
+                            )
+                        else:
+                            self.assertFalse(updated)
+                            alarm.assert_equal(acked_alarm)
+
+            self.assertEqual(self.ncalls, desired_ncalls)
 
     async def test_unacknowledge(self):
         user = "skipper"
@@ -604,14 +732,14 @@ class AlarmTestCase(asynctest.TestCase):
 
             # unacknowledge should have no effect initially
             # because alarm is not acknowledged
-            alarm = copy.copy(alarm0)
+            alarm = self.copy_alarm(alarm0)
             updated = alarm.unacknowledge()
             self.assertFalse(updated)
-            self.assertEqual(alarm, alarm0)
+            alarm.assert_equal(alarm0)
 
             # acknowledge the alarm
             for ack_severity in AlarmSeverity:
-                alarm = copy.copy(alarm0)
+                alarm = self.copy_alarm(alarm0)
                 if ack_severity < alarm.max_severity:
                     continue
                 updated = alarm.acknowledge(severity=ack_severity, user=user)
@@ -624,12 +752,12 @@ class AlarmTestCase(asynctest.TestCase):
                     self.assertFalse(alarm.nominal)
 
                 # unacknowledge the alarm
-                acked_alarm = copy.copy(alarm)
+                acked_alarm = self.copy_alarm(alarm)
                 tai0 = salobj.current_tai()
                 updated = alarm.unacknowledge()
                 if acked_alarm.nominal:
                     self.assertFalse(updated)
-                    self.assertEqual(alarm, acked_alarm)
+                    alarm.assert_equal(acked_alarm)
                 else:
                     desired_ncalls += 1
                     self.assertTrue(updated)
@@ -650,15 +778,38 @@ class AlarmTestCase(asynctest.TestCase):
 
         self.assertEqual(self.ncalls, desired_ncalls)
 
+    async def test_escalate(self):
+        escalate_delay = 0.1
+        escalate_to = "chaos"
+
+        for alarm in self.alarm_iter(
+            name="user",
+            callback=self.callback,
+            escalate_delay=escalate_delay,
+            escalate_to=escalate_to,
+        ):
+            self.assertFalse(alarm.escalated)
+            self.assertEqual(alarm.escalate_delay, escalate_delay)
+            self.assertEqual(alarm.escalate_to, escalate_to)
+            if alarm.max_severity < AlarmSeverity.CRITICAL:
+                self.assertEqual(alarm.timestamp_escalate, 0)
+                self.assertTrue(alarm.escalate_task.done())
+            else:
+                self.assertGreater(alarm.timestamp_escalate, 0)
+                self.assertFalse(alarm.escalate_task.done())
+                await self.next_queued_alarm(expected_alarm=alarm)
+                self.assertTrue(alarm.escalated)
+
     async def test_reset(self):
         name = "alarm"
         blank_alarm = self.make_alarm(name=name, callback=None)
         blank_alarm.callback = self.callback
+        self.assertTrue(blank_alarm.nominal)
         for alarm in self.alarm_iter(name=name, callback=self.callback):
             if not alarm.nominal:
                 self.assertNotEqual(alarm, blank_alarm)
             alarm.reset()
-            self.assertEqual(alarm, blank_alarm)
+            alarm.assert_equal(blank_alarm)
 
     async def test_mute_valid(self):
         user = "otho"
@@ -666,13 +817,12 @@ class AlarmTestCase(asynctest.TestCase):
         for severity in AlarmSeverity:
             if severity == AlarmSeverity.NONE:
                 continue  # invalid value
-            callback = AsyncCallback()
-            for alarm in self.alarm_iter(name=user, callback=callback):
+            for alarm in self.alarm_iter(name=user, callback=self.callback):
                 t0 = salobj.current_tai()
                 alarm.mute(duration=duration, severity=severity, user=user)
                 curr_tai = salobj.current_tai()
                 dt = curr_tai - t0
-                await callback.next(timeout=STD_TIMEOUT)
+                await self.next_queued_alarm(expected_alarm=alarm)
                 self.assertTrue(alarm.muted)
                 self.assertEqual(alarm.muted_by, user)
                 self.assertEqual(alarm.muted_severity, severity)
@@ -683,7 +833,9 @@ class AlarmTestCase(asynctest.TestCase):
                     alarm.timestamp_unmute, curr_tai + duration, delta=dt
                 )
                 # Wait for the alrm to unmute itself.
-                await callback.next(timeout=STD_TIMEOUT + duration)
+                await self.next_queued_alarm(
+                    expected_alarm=alarm, timeout=STD_TIMEOUT + duration
+                )
                 self.assertFalse(alarm.muted)
                 self.assertEqual(alarm.muted_by, "")
                 self.assertEqual(alarm.muted_severity, AlarmSeverity.NONE)
@@ -700,7 +852,7 @@ class AlarmTestCase(asynctest.TestCase):
             ):
                 # check that mute raises ValueError for invalid values
                 # and leaves the alarm state unchanged
-                initial_alarm = copy.copy(alarm)
+                initial_alarm = self.copy_alarm(alarm)
                 with self.assertRaises(ValueError):
                     alarm.mute(
                         duration=bad_delay, severity=good_severity, user=failed_user
@@ -713,32 +865,32 @@ class AlarmTestCase(asynctest.TestCase):
                     alarm.mute(
                         duration=bad_delay, severity=bad_severity, user=failed_user
                     )
-                self.assertEqual(alarm, initial_alarm)
+                alarm.assert_equal(initial_alarm)
 
                 # make sure failures also leave muted alarm state unchanged
                 alarm.mute(duration=good_delay, severity=good_severity, user=good_user)
                 self.assertTrue(alarm.muted)
                 self.assertEqual(alarm.muted_by, good_user)
                 self.assertEqual(alarm.muted_severity, good_severity)
-                muted_alarm = copy.copy(alarm)
+                muted_alarm = self.copy_alarm(alarm)
 
                 with self.assertRaises(ValueError):
                     alarm.mute(
                         duration=bad_delay, severity=good_severity, user=failed_user
                     )
-                self.assertEqual(alarm, muted_alarm)
+                alarm.assert_equal(muted_alarm)
 
                 with self.assertRaises(ValueError):
                     alarm.mute(
                         duration=good_delay, severity=bad_severity, user=failed_user
                     )
-                self.assertEqual(alarm, muted_alarm)
+                alarm.assert_equal(muted_alarm)
 
                 with self.assertRaises(ValueError):
                     alarm.mute(
                         duration=bad_delay, severity=bad_severity, user=failed_user
                     )
-                self.assertEqual(alarm, muted_alarm)
+                alarm.assert_equal(muted_alarm)
 
                 alarm.unmute()  # kill unmute timer
 
@@ -749,19 +901,19 @@ class AlarmTestCase(asynctest.TestCase):
             if severity == AlarmSeverity.NONE:
                 continue  # invalid value
             for alarm in self.alarm_iter(name=user, callback=self.callback):
-                self.ncalls = 0
+                ncalls0 = self.ncalls
                 # check that unmute on unmuted alarm is a no-op
-                original_alarm = copy.copy(alarm)
+                original_alarm = self.copy_alarm(alarm)
                 alarm.unmute()
-                self.assertEqual(alarm, original_alarm)
-                self.assertEqual(self.ncalls, 1)
+                alarm.assert_equal(original_alarm)
+                self.assertEqual(self.ncalls, ncalls0 + 1)
 
                 # mute alarm and unmute it again before it unmutes itself
                 t0 = salobj.current_tai()
                 alarm.mute(duration=duration, severity=severity, user=user)
                 curr_tai = salobj.current_tai()
                 dt = curr_tai - t0
-                self.assertEqual(self.ncalls, 2)
+                self.assertEqual(self.ncalls, ncalls0 + 2)
                 self.assertTrue(alarm.muted)
                 self.assertEqual(alarm.muted_by, user)
                 self.assertEqual(alarm.muted_severity, severity)
@@ -771,12 +923,12 @@ class AlarmTestCase(asynctest.TestCase):
                 )
 
                 alarm.unmute()
-                self.assertEqual(self.ncalls, 3)
+                self.assertEqual(self.ncalls, ncalls0 + 3)
                 # Give asyncio a chance to cancel the mute task.
                 await asyncio.sleep(0)
                 self.assertTrue(alarm.unmute_task.done())
                 # Compare equality.
-                self.assertEqual(alarm, original_alarm)
+                alarm.assert_equal(original_alarm)
 
     async def test_repr(self):
         name = "Something.else"
@@ -794,7 +946,7 @@ class AlarmTestCase(asynctest.TestCase):
 
             # acknowledge the alarm
             for ack_severity in AlarmSeverity:
-                alarm = copy.copy(alarm0)
+                alarm = self.copy_alarm(alarm0)
                 if ack_severity < alarm.max_severity:
                     continue
                 updated = alarm.acknowledge(severity=ack_severity, user=user)
@@ -808,7 +960,7 @@ class AlarmTestCase(asynctest.TestCase):
 
                 acked_alarm = alarm
                 for severity in AlarmSeverity:
-                    alarm = copy.copy(acked_alarm)
+                    alarm = self.copy_alarm(acked_alarm)
                     tai0 = salobj.current_tai()
                     reason = f"set severity to {severity} after ack"
                     updated = alarm.set_severity(severity, reason=reason)
@@ -828,7 +980,7 @@ class AlarmTestCase(asynctest.TestCase):
                     if severity == AlarmSeverity.NONE:
                         if acked_alarm.nominal:
                             self.assertFalse(updated)
-                            self.assertEqual(alarm, acked_alarm)
+                            alarm.assert_equal(acked_alarm)
                         else:
                             # alarm should be reset
                             self.assertTrue(updated)